chore: clean up all remaining clippy warnings workspace-wide
Final batch: bitvue-av3-codec, bitvue-avs3, bitvue-vc3, bitvue-vvc, bitvue-vp9,
bitvue-hevc, bitvue-formats, bitvue-codecs-parser, bitvue-metrics, bitvue-engine.
Combined with the three prior commits (av1-codec, cli, avc+mpeg2-codec+decode),
this closes out the workspace-wide clippy cleanup -- `cargo clippy --workspace
--lib --tests` now shows zero warnings anywhere outside crates/vendor/abseil
(third-party vendored code, explicitly out of scope, left untouched).

Several crates had `#[deny(clippy::absurd_extreme_comparisons)]` violations
that were hard compile errors, not just warnings -- these had been silently
preventing `cargo test`/`cargo clippy --tests` from running at all for
bitvue-vvc, bitvue-vp9, bitvue-hevc, and bitvue-engine until fixed here.

Real bugs found while investigating tautological-assertion warnings (the same
category already fixed several times elsewhere this session):
- bitvue-vp9/tests/functional_coverage_test.rs: two tests built VP9 fixtures
  with an invalid sync-code byte (0x00 instead of the required 0x42), so
  parse_vp9()/extract_vp9_frames() always produced zero frames -- a
  tautological `>= 0` assertion was masking this. Fixed to assert the real
  (0) value with an explanation, rather than silently "fixing" the sync code
  and changing what the test covers.
- bitvue-vp9/tests/vp9_stress_test.rs: test_parse_vp9_all_frame_types looped
  over frame_type in [0,1,2] but hardcoded `(0 << 3)` in the byte
  construction instead of using the loop variable -- all 3 iterations tested
  identical bytes. Fixed to use the loop variable.
- bitvue-vvc/tests/vvc_edge_cases_test.rs: traced find_nal_units's byte-offset
  logic by hand to confirm a zero-length-NAL-span edge case really does
  produce 0 NAL units, replacing a vacuous `>= 0` check with the real value.
- bitvue-av3-codec/tests/integration_test.rs: test_av3_obu_types had OBU type
  byte values that never matched the real ObuType enum encoding, so a
  tautological assert!(true) inside a rarely-true branch almost never even
  executed -- fixed the encoding and asserted real values.
- bitvue-engine/src/core_test.rs: `let _ = selection.read()` (let_underscore_lock)
  investigated and confirmed benign (only verifies the read-lock acquisition
  doesn't panic, nothing downstream depends on holding it) -- given a named
  `let _guard = ...` binding rather than treated as a hidden bug.

Real bitstream-parsing files (bitvue-av3-codec/avs3/vc3/vvc/vp9/hevc's
overlay_extraction.rs/bitreader.rs/frame_header.rs, bitvue-formats/ts.rs) got
only syntax-preserving mechanical fixes, each confirmed against the
surrounding function's real logic before editing -- no parsing-behavior
changes. bitvue-engine/src/diagnostics_bands.rs's one "manual checked
division" warning was fixed with checked_div(), verifying the existing
divide-by-zero behavior is preserved.

cargo build --workspace clean. cargo test --workspace --lib --tests: 141 test
binaries, 0 failures (the one known pre-existing flaky test,
bitvue-engine::compare_cache's LRU eviction test under parallel execution,
is untouched and unrelated -- confirmed passing 100% under --test-threads=1).
cargo clippy --workspace --lib --tests: 0 non-vendor warnings, confirmed by
grepping every `-->` path in the output. cargo fmt --all -- --check: clean.

+++ crates/bitvue-vp9/src/overlay_extraction.rs
@@ -465,7 +465,7 @@ fn parse_super_blocks_inner(
         let qp = if seg.enabled && seg.feature_enabled[seg_id as usize][0] {
             let alt_q = seg.feature_data[seg_id as usize][0];
             if seg.abs_or_delta_update {
-                alt_q.clamp(0, 255) as i16
+                alt_q.clamp(0, 255)
             } else {
                 (base_qp + alt_q).clamp(0, 255)
             }
